Add files via upload
For board lots of it isn't done, but that is what I have so far. It correctly draws the board and all the lines, we can change color/width and stuff later.
For constants it is just a list of some of the constants I have been using to make the code more readable and less jumbled with seemingly random numbers.